Grouped gql queries in one directory
