+++ components/account/AccountPage.tsx
@@ -33,16 +33,9 @@ import ContentLoader from "react-content-loader";
 import numbro from "numbro";
 import AccountOverview from "./AccountOverview";
 import CoinBalanceHistory from './CoinBalanceHistory'
+import { GET_ACCOUNT_DETAILS } from './../query/Account'
 
-const GET_ACCOUNT_DETAILS = gql`
-  query Account($address: String!) {
-    account(address: $address) {
-      _id
-      address
-      balance
-    }
-  }
-`;
+
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
