+++ components/block/LatestBlocks.tsx
@@ -26,26 +26,7 @@ import numbro from "numbro";
 import { useRouter } from "next/router";
 
 import MiddleEllipsis from './../MiddleEllipsis'
-
-const GET_BLOCK = gql`
-  {
-    blocks(pageSize: 500, page: 10) {
-      blocks {
-        number
-        miner {
-          name
-          affiliation
-        }
-        transactions {
-          transactionIndex
-        }
-        gasUsed
-        gasLimit
-        timestamp
-      }
-    }
-  }
-`;
+import { GET_BLOCK } from './../query/Block'
