+++ components/proposal/ProposalDetails.tsx
@@ -16,27 +16,6 @@ import ContentLoader from "react-content-loader";
 import moment from "moment";
 import LedgerButtons from "../ledger/LedgerButtons";
 
-// const GET_BLOCK_DETAILS = gql`
-//   query Block($number: Int) {
-//     block(number: $number) {
-//       timestamp
-//       transactions {
-//         transactionIndex
-//         nonce
-//       }
-//       size
-//       miner {
-//         name
-//         signer
-//       }
-//       hash
-//       parentHash
-//       totalDifficulty
-//       gasUsed
-//       gasLimit
-//     }
-//   }
-// `;
 
 const useStyles = makeStyles(() => {
   return {
