+++ components/query/Chain.tsx
@@ -1,0 +1,22 @@
+import React from "react";
+import { gql } from "apollo-boost";
+import { useQuery } from "@apollo/react-hooks";
+
+
+export const GET_CHAIN = gql`
+  {
+    chain {
+      _id
+      averageBlockTime
+      txCount
+      latestHeight
+      chainId
+      tokenPrice {
+        usd
+        usdMarketCap
+      }
+      walletCount
+    }
+  }
+`;
+
